Update Makefile all target and remove faulty test
Modify the 'all' target in Makefile to build both the main application and the test runner by default. Remove the unused #include for assert.h and the entire test_sa_over_allocation function from test_stack_alloc.c, as it was testing over-allocation behavior that could be incorrect or redundant. This cleans up the build process and test suite.